pyproject.toml
[tool.poetry]
name = "delivery-project"
version = "0.1.0"
description = ""
authors = ["regvud <silveron.regvud@gmail.com>"]
readme = "README.md"
packages = [{include = "delivery_project"}]

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework = "^3.14.0"
mysqlclient = "^2.2.0"
django-dotenv = "^1.4.2"
djangorestframework-simplejwt = "^5.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
